Release 1.5 * refactoring + notification

+++ create_archive.sh
@@ -1,27 +1,21 @@
 #!/bin/bash
 
-cd sources
-
-FILES="CMakeLists.txt configwindow.ui oblikue-strategies.cpp oblikue-strategies.h oblikue-strategies.png plasma-applet-oblikue-strategies.desktop ../README"
 ARCHIVE="oblikue-strategies"
-VERSION=`cat plasma-applet-oblikue-strategies.desktop| grep Version | awk -F "=" '{print $2}'`
+SRCDIR="sources"
+FILES="README"
+IGNORELIST="CMakeLists.txt.user"
+VERSION=`grep Version ${SRCDIR}/*.desktop | awk -F "=" '{print $2}'`
 
 # create archive
-if [ -e ${ARCHIVE}-${VERSION}.tar.xz ]; then
-  rm -f ${ARCHIVE}-${VERSION}.tar.xz
-fi
-if [ -d ${ARCHIVE} ]; then
-  rm -rf ${ARCHIVE}
-fi
-mkdir ${ARCHIVE}
-for FILE in ${FILES[*]}; do
-  cp $FILE ${ARCHIVE}
-done
-tar cvJf ../${ARCHIVE}-${VERSION}.tar.xz ${ARCHIVE}
+[[ -e ${ARCHIVE}-${VERSION}.tar.xz ]] && rm -f ${ARCHIVE}-${VERSION}.tar.xz
+[[ -d ${ARCHIVE} ]] && rm -rf ${ARCHIVE}
+cp -r ${SRCDIR} $ARCHIVE
+for FILE in ${FILES[*]}; do cp $FILE ${ARCHIVE}; done
+for FILE in ${IGNORELIST[*]}; do rm -f ${ARCHIVE}/${FILE}; done
+tar cvJf ${ARCHIVE}-${VERSION}.tar.xz ${ARCHIVE} > /dev/null
 rm -rf ${ARCHIVE}
 
 # update md5sum
-cd ..
-MD5SUMS=`md5sum ${ARCHIVE}-${VERSION}.tar.xz| awk '{print $1}'`
+MD5SUMS=`md5sum ${ARCHIVE}-${VERSION}.tar.xz | awk '{print $1}'`
 sed -i "/md5sums=('[0-9A-Fa-f]*/s/[^'][^)]*/md5sums=('${MD5SUMS}'/" PKGBUILD
 sed -i "s/pkgver=[0-9.]*/pkgver=${VERSION}/" PKGBUILD
